split out snapshots, recording downloads, archived recordings
